Second project of the course
Basics of ROS: nodes, services, publishers, subscribers, etc.
In this project, a two-degree of freedom mobile robot with a LiDAR sensor and a camera is created. Using ROS nodes and services, the robot can process raw images from a camera and detect a white ball in the environment and follow it using a simple control algorithm.